WIP - Move refs into separate data file
We have some circular dependencies related to the Modal which seem to be
related to the refs.

[skip ci]

Jira ticket: CAMS-47
