+++ user-interface/src/components/CaseAssignment.tsx
@@ -3,9 +3,9 @@ import { useState, useEffect, useRef } from 'react';
 import Api from '../models/api';
 import { Chapter15Type, Chapter15CaseListResponseData } from '../type-declarations/chapter-15';
 import MockApi from '../models/chapter15-mock.api.cases';
-import { ToggleModalButton } from './uswds/ToggleModalButton';
+import { ToggleModalButton } from './uswds/modal/ToggleModalButton';
 import AssignAttorneyModal, { CallBackProps } from './AssignAttorneyModal';
-import { ModalRefType } from './uswds/Modal';
+import { ModalRefType } from './uswds/modal/Modal';
 import Alert, { AlertRefType, UswdsAlertStyle } from './uswds/Alert';
 import AttorneysApi from '../models/attorneys-api';
 import { Attorney } from '../type-declarations/attorneys';
