+++ user-interface/src/components/uswds/modal/modal-refs.d.ts
@@ -1,0 +1,12 @@
+import { RefObject } from 'react';
+import { SubmitCancelButtonGroupRef } from './SubmitCancelButtonGroup';
+
+export interface ModalRefType {
+  show: () => void;
+  hide: () => void;
+  buttons?: RefObject<SubmitCancelButtonGroupRef>;
+}
+
+export interface SubmitCancelButtonGroupRef {
+  disableSubmitButton: (state: boolean) => void;
+}
